Develop

+++ src/Controller/Api/WalkController.php
@@ -10,9 +10,6 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
-
-
 
 
 class WalkController extends AbstractController
@@ -58,11 +55,10 @@ class WalkController extends AbstractController
 
     /**
      * Delete a walk
-     * @Route("/api/walks/delete/{id<\d+>}", name="api_walks_delete_item", methods={"DELETE"})
+     * @Route("/api/walks/{id<\d+>}", name="api_walks_delete_item", methods={"DELETE"})
      */
-    public function deleteItem(Walk $walk = null, EntityManagerInterface $em, TokenInterface $token)
+    public function deleteItem(Walk $walk = null, EntityManagerInterface $em)
     {
-        
         // managing error
         if ($walk === null) {
 
@@ -76,7 +72,7 @@ class WalkController extends AbstractController
             return $this->json($message, Response::HTTP_NOT_FOUND);
         }
 
-    
+        $this->denyAccessUnlessGranted('delete', $walk);
     
         // Delete
         $em->remove($walk);
